Metoder flyttet rundt til de rigtige lag
