+++ src/logic/AllCarModels.java
@@ -1,13 +1,11 @@
 package logic;
 
-import data.CarModel;
-
 import java.util.ArrayList;
 
 public class AllCarModels {
-    public static ArrayList<CarModel> allCarModels = new ArrayList<>();
+    public static ArrayList<Car> allCars = new ArrayList<>();
 
-    public void addTeam(CarModel carModel){
-        allCarModels.add(carModel);
+    public void addCar(Car car){
+        allCars.add(car);
     }
 }
